LPD-5579 portal-search-web: Refactor facet util aui component
