+++ modules/apps/portal-search/portal-search-web/src/main/resources/META-INF/resources/custom/facet/view.jsp
@@ -12,6 +12,7 @@
 <%@ taglib uri="http://liferay.com/tld/aui" prefix="aui" %><%@
 taglib uri="http://liferay.com/tld/clay" prefix="clay" %><%@
 taglib uri="http://liferay.com/tld/ddm" prefix="liferay-ddm" %><%@
+taglib uri="http://liferay.com/tld/frontend" prefix="liferay-frontend" %><%@
 taglib uri="http://liferay.com/tld/ui" prefix="liferay-ui" %>
 
 <%@ page import="com.liferay.petra.string.StringPool" %><%@
@@ -128,12 +129,11 @@ CustomFacetPortletInstanceConfiguration customFacetPortletInstanceConfiguration 
 	</c:otherwise>
 </c:choose>
 
-<aui:script use="liferay-search-facet-util">
-	var facetTerms = document.querySelectorAll(
-		'#<portlet:namespace />fm .facet-term'
-	);
-
-	facetTerms.forEach((term) => {
-		Liferay.Util.toggleDisabled(term, false);
-	});
-</aui:script>
+<liferay-frontend:component
+	context='<%=
+		HashMapBuilder.<String, Object>put(
+			"namespace", liferayPortletResponse.getNamespace()
+		).build()
+	%>'
+	module="js/FacetUtil"
+/>
