+++ modules/apps/portal-search/portal-search-web/src/main/resources/META-INF/resources/js/custom_filter.js
@@ -94,6 +94,6 @@ AUI.add(
 	},
 	'',
 	{
-		requires: ['liferay-search-facet-util'],
+		requires: [],
 	}
 );
